mesh-adoption/quality/cdp-quality-sidecar/requirements.txt
# great_expectations==1.2.6
# numpy
# pandas
# pyodbc
# certifi==2024.8.30
# charset-normalizer==3.4.0
# Deprecated==1.2.15
# fire==0.7.0
# googleapis-common-protos==1.66.0
# grpcio==1.68.0
# idna==3.10
# importlib_metadata==8.5.0
# opentelemetry-api==1.28.2
# opentelemetry-distro==0.49b2
# opentelemetry-exporter-otlp==1.28.2
# opentelemetry-exporter-otlp-proto-common==1.28.2
# opentelemetry-exporter-otlp-proto-grpc==1.28.2
# opentelemetry-exporter-otlp-proto-http==1.28.2
# opentelemetry-proto==1.28.2
# opentelemetry-sdk==1.28.2
# opentelemetry-semantic-conventions==0.49b2
# opentelemetry-util-http==0.49b2
# packaging==24.2
# protobuf==5.28.3
# randomname==0.2.1
# requests==2.32.3
# termcolor==2.5.0
# typing_extensions==4.12.2
# urllib3==2.2.3
# wrapt==1.16.0
# zipp==3.21.0

great_expectations==1.3.0
pandas
opentelemetry-api==1.28.2
opentelemetry-distro==0.49b2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
wrapt==1.16.0
